requirements.txt
google-genai>=1.1.0
pyaudio>=0.2.13
python-dotenv